packages/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Evergreen Multi Agents API Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.4",
    "google-genai>=1.56.0",
    "pydantic-settings>=2.12.0",
    "google-adk>=1.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
